Streamlit/requirements.txt
streamlit
pandas
numpy
streamlit-authenticator
google-cloud-bigquery
google-cloud-storage